% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_screen}
\alias{bin_screen}
\title{Binary IRT Screener}
\usage{
bin_screen(X, Dim = 2, thresh = 0.35, control = NULL)
}
\arguments{
\item{X}{Roll call matrix}

\item{Dim}{Number of dimensions to estimate}

\item{thresh}{Threshold for the K-means clustering in the final
stage.  If all PRE values are above \code{thresh}, then no clustering will
be done and all and all remaining variables will be retained.}

\item{control}{List of control arguments for the \code{binIRT()} function from the \code{emIRT} package.}
}
\description{
Implements a screening pass of a cross-sectional binary IRT to identify
variables to be included for each dimension.
}
