% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster.pre}
\alias{cluster.pre}
\title{Proportional Reduction in Error for Binary IRT Models}
\usage{
cluster.pre(obj, data, threshold)
}
\arguments{
\item{obj}{Object of class \code{binIRT}.}

\item{data}{The data used to estimate \code{obj}.}

\item{threshold}{Threshold for PRE values - if all
PRE values are greater than the threshold, all
variables will be identified as in the same cluster.}
}
\value{
A list with the group for each variable in
\code{cluster} and the cluster number with the high PRE
values in \code{high.clust}.
}
\description{
Calculates the proportional reduction in error for
binary IRT models estimated with \code{binIRT()} from the
\code{emIRT} package. `
}
