% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seq_dynIRT}
\alias{seq_dynIRT}
\title{Sequential Dynamic MIRT Estimation}
\usage{
seq_dynIRT(.data, Dim, thresh = 0.35, ...)
}
\arguments{
\item{.data}{Data frame with only variables that
get passed to the IRT function.}

\item{Dim}{Number of dimensions to estimate}

\item{thresh}{Threshold for the K-means clustering in the final
stage.  If all PRE values are above \code{thresh}, then no clustering will
be done and all and all remaining variables will be retained.}

\item{...}{Arguments to be passed down to the \code{binIRT}
function including \code{.data}, \code{.starts} and \code{.priors}.}
}
\value{
A list with the output from each final \code{binIRT} model.
}
\description{
Estimates a multidimensional dynamic IRT model by sequentially
identifying variables that correspond with each dimension.
Once identified, the models  are re-estimated with only the
relevant variables.
}
